requirements.txt
# Prefer binary wheels (avoids building numpy on Python 3.13 / Windows)
numpy>=2.0.0

fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic>=2.0.0

langchain>=0.2.16
langchain-core>=0.2.39
langchain-community>=0.2.16
langchain-openai>=0.1.23
langchain-text-splitters>=0.2.4
langchain-qdrant>=0.1.4
langgraph>=0.2.35

qdrant-client>=1.11.1
pypdf>=4.3.1
tavily-python>=0.5.0
rank_bm25>=0.2.2

# Eval + optional
ragas>=0.2.0
datasets>=2.14.0
langsmith>=0.1.0
langchain-cohere>=0.0.20
langchain-experimental>=0.3.0
